import disk-io settings

+++ src/disk_io_thread.cpp
@@ -1573,7 +1573,7 @@ namespace libed2k
 			int unchoke_limit = m_settings.unchoke_slots_limit;
 			if (unchoke_limit < 0) unchoke_limit = 100;
 
-			if (m_sorted_read_jobs.size() > unchoke_limit * 2)
+			if ((int)m_sorted_read_jobs.size() > unchoke_limit * 2)
 			{
 				int range = unchoke_limit;
 				int exceed = m_sorted_read_jobs.size() - range * 2;
